requirements.txt
streamlit>=1.30.0
pytesseract>=0.3.10
Pillow>=10.0.0
rapidfuzz>=3.5.0
ollama>=0.1.0
